# Merge optimizer 5319