Regarding issue #5319 , I changed the broadcastable pattern in case of TypeError exception. In an a special case that ```dtypes, ndims``` match but the broadcastable pattern doesn't, we will try to change the broadcastable pattern of one of them to be compatible with the other one. We first check for these conditions and then call the ```fgraph.replace_all_validate``` function.