@nouiz 2017-02-03
Can you confirm that the error goes away with this version of the fix?