@ReyhaneAskari 2017-02-06
Dear Fred, @nouiz 

This fixes the error which was previously generated. All the tests for GPU and CPU are passed. DLT tests also passed. The result for benchmarking with SBRNN (29 layers) is as follows:

**Master:**
  Time in 352 calls to Function.__call__: 2.373188e+02s
  Time in Function.fn.__call__: 2.368799e+02s (99.815%)
  Time in thunks: 2.099577e+02s (88.471%)
  Total compile time: **1.489671e+03s**

2.579713s - ('merge1', 'MergeOptimizer', 0, 39948, 28372) - **0.124s**
     MergeOptimizer
       nb fail=    0 merged=17770 constant= 6194
       time replace=1.76 validate=0.12 callback=1.06

2.033230s - ('merge2', 'MergeOptimizer', 23, 21157, 15659) - **0.182s**
     MergeOptimizer
       nb fail=    0 merged=22595 constant=11281
       time replace=2.03 validate=0.18 callback=1.20

29.190615s - ('merge3', 'MergeOptimizer', 59, 15659, 15659) - **29.141s**
     MergeOptimizer
       nb fail=    0 merged=  124 constant=  124
       time replace=29.19 validate=29.14 callback=29.18

-------------------------------------
**New Commit:** 
  Time in 352 calls to Function.__call__: 2.373914e+02s
  Time in Function.fn.__call__: 2.369307e+02s (99.806%)
  Time in thunks: 2.102712e+02s (88.576%)
  Total compile time: **1.679333e+03s**

3.293359s - ('merge1', 'MergeOptimizer', 0, 39948, 28372) - **0.123s**
     MergeOptimizer
       nb fail=    0 merged=17770 constant= 6194
       time replace=1.73 validate=0.12 callback=1.00

2.071804s - ('merge2', 'MergeOptimizer', 23, 21157, 15659) - **0.180s**
     MergeOptimizer
       nb fail=    0 merged=22570 constant=11256
       time replace=2.07 validate=0.18 callback=1.20

34.473688s - ('merge3', 'MergeOptimizer', 59, 15659, 15659) - **34.412s**
     MergeOptimizer
       nb fail=    0 merged=  149 constant=  149
       time replace=34.47 validate=34.41 callback=34.46