@ReyhaneAskari 2017-02-06
Hi Fred,

Attached is the full profile of both runs. A more complete version of merge3 follows:

**Master:**
29.190615s - ('**merge3**', 'MergeOptimizer', 59, 15659, 15659) - 29.141s
     MergeOptimizer
       nb fail=    0 merged=  124 constant=  124
       time replace=29.19 validate=29.14 callback=29.18
       callbacks_time
            <theano.gof.toolbox.PreserveVariableAttributes object at 0x7f502d4c0b10> , 0.000388383865356
            <theano.gof.toolbox.ReplaceValidate object at 0x7f502dfe5a50> , 0.000644445419312
            <theano.gof.opt.MergeFeature object at 0x7f502c4025d0> , 0.00394439697266
            <theano.tensor.opt.ShapeFeature object at 0x7f502c2ad050> , 0.0282342433929
            <theano.compile.function_module.Supervisor instance at 0x7f5028c3fd40> , 3.7806596756
            **<theano.gof.destroyhandler.DestroyHandler object at 0x7f50106ffad0> , 25.3625121117**

**New Commit:**
34.473688s - ('**merge3**', 'MergeOptimizer', 59, 15659, 15659) - 34.412s
     MergeOptimizer
       nb fail=    0 merged=  149 constant=  149
       time replace=34.47 validate=34.41 callback=34.46
       callbacks_time
            <theano.gof.toolbox.PreserveVariableAttributes object at 0x7faffe9c4e50> , 0.000468015670776
            <theano.gof.toolbox.ReplaceValidate object at 0x7fafff511a50> , 0.000803232192993
            <theano.gof.opt.MergeFeature object at 0x7faffdd7dd50> , 0.00532221794128
            <theano.tensor.opt.ShapeFeature object at 0x7faffb76b490> , 0.0312221050262
            <theano.compile.function_module.Supervisor instance at 0x7faffa1c9c20> , 2.33457803726
            **<theano.gof.destroyhandler.DestroyHandler object at 0x7fafe2302710> , 32.0809686184**

It seems that the function which takes most of the time is the DestroyHandler. It seems to have taken a lot of time in the overall callbacks for the new Commit too:

**Master:**

Optimizer Profile
 SeqOptimizer  OPT_FAST_RUN  time 8.257s for 1561/1743 nodes before/after optimization
   1.011s for callback
       0.330s for fgraph.validate()
   callbacks_time
        <theano.tensor.opt.ShapeFeature object at 0x7f4fcf1af5d0> , 0.431351423264
        **<theano.gof.destroyhandler.DestroyHandler object at 0x7f4fcd541490> , 0.302550077438**
        <theano.gof.opt.MergeFeature object at 0x7f4fcf79fbd0> , 0.141258955002
        <theano.compile.function_module.Supervisor instance at 0x7f4fcee7cb00> , 0.0296409130096


**New Commit:**

Optimizer Profile
 SeqOptimizer  OPT_FAST_RUN  time 484.438s for 39948/15659 nodes before/after optimization
   185.688s for callback
       158.478s for fgraph.validate()
   callbacks_time
        **<theano.gof.destroyhandler.DestroyHandler object at 0x7fafe2302710> , 146.390594482**
        <theano.compile.function_module.Supervisor instance at 0x7faffa1c9c20> , 22.370765686
        <theano.tensor.opt.ShapeFeature object at 0x7faffb76b490> , 17.5543973446
        <theano.gof.opt.MergeFeature object at 0x7faffdd7dd50> , 6.49993062019

[master_5319.txt](https://github.com/Theano/Theano/files/755338/master_5319.txt)
[new_commit_5319.txt](https://github.com/Theano/Theano/files/755337/new_commit_5319.txt)