@ReyhaneAskari 2017-02-10
@nouiz , I tried to implement your suggestion to find out which nodes were being merged in ```merge3```. It was not successful as the dictionary was always empty except for once. So during the whole execution for we only had one Constant pairs named ```merge3```. I tried to trace this though the profiling and see where this  line: 

```29.661805s - ('merge3', 'MergeOptimizer', 60, 15659, 15659) - 29.610s```  

was being printed and how the nodes were counted. I tried couple of other approaches as well but was not successful.

Finally, I replaced ```dict```  to ```OrderedDict```  inside  ```sb_resnet.py``` and the problem was fixed. Now each time we run the benchmark we get the same number of merged nodes for all the merges. 

Following is the result of benchmarking. As it can be seen, there difference between the versions is really not significant. 

**Master:** [old_master.txt](https://github.com/Theano/Theano/files/767806/old_master.txt)
0.538894s - ('**merge1**', 'MergeOptimizer', 0, 12092, 8232) - **0.039s**
     MergeOptimizer
       nb fail=    0 merged= **5604** constant= **1744**
       time replace=0.35 validate=0.04 callback=0.19

0.348806s - ('**merge2**', 'MergeOptimizer', 23, 5159, 3775) - **0.045s**
     MergeOptimizer
       nb fail=    0 merged= **5755** constant= **2617**
       time replace=0.35 validate=0.05 callback=0.21

5.030370s - ('**merge3**', 'MergeOptimizer', 51, 3775, 3775) - **5.014s**
     MergeOptimizer
       nb fail=    0 merged=  **106** constant=  **106**
       time replace=5.03 validate=5.01 callback=5.02

------------------------------------------------------------------------

**New Commit:** [newcommit.txt](https://github.com/Theano/Theano/files/767805/newcommit.txt)
0.588227s - ('**merge1**', 'MergeOptimizer', 0, 12092, 8232) - **0.040s**
     MergeOptimizer
       nb fail=    0 merged= **5604** constant= **1744**
       time replace=0.37 validate=0.04 callback=0.19

0.356588s - ('**merge2**', 'MergeOptimizer', 23, 5159, 3775) - **0.045s**
     MergeOptimizer
       nb fail=    0 merged= **5755** constant= **2617**
       time replace=0.36 validate=0.05 callback=0.21

5.294038s - ('**merge3**', 'MergeOptimizer', 51, 3775, 3775) - **5.275s**
     MergeOptimizer
       nb fail=    0 merged=  **106** constant=  **106**
       time replace=5.29 validate=5.28 callback=5.29